+++ src/grammar/GrammarCustomConstructorsNode.ts
@@ -91,15 +91,56 @@ class CustomBrowserConstructorNode extends AbstractCustomConstructorNode {
   }
 }
 
+class CustomJavascriptConstructorNode extends AbstractCustomConstructorNode {
+  private _cached: any
+  static cache: any = {}
+  private _getNodeJsConstructor(): types.RunTimeNodeConstructor {
+    const jtreePath = __dirname + "/../jtree.node.js"
+    const code = `const jtree = require('${jtreePath}').default
+/* INDENT FOR BUILD REASONS */  module.exports = ${this.childrenToString()}`
+    if (CustomJavascriptConstructorNode.cache[code]) return CustomJavascriptConstructorNode.cache[code]
+    const tempFilePath = __dirname + "/constructor-" + TreeUtils.getRandomString(30) + "-temp.js"
+    const fs = require("fs")
+    try {
+      fs.writeFileSync(tempFilePath, code, "utf8")
+      CustomJavascriptConstructorNode.cache[code] = require(tempFilePath)
+    } catch (err) {
+      console.error(err)
+    } finally {
+      fs.unlinkSync(tempFilePath)
+    }
+
+    return CustomJavascriptConstructorNode.cache[code]
+  }
+
+  private _getBrowserConstructor(): types.RunTimeNodeConstructor {
+    const definedCode = this.childrenToString()
+    const tempClassName = "tempConstructor" + TreeUtils.getRandomString(30)
+    if (CustomJavascriptConstructorNode.cache[definedCode]) return CustomJavascriptConstructorNode.cache[definedCode]
+
+    const script = document.createElement("script")
+    script.innerHTML = `window.${tempClassName} = ${this.childrenToString()}`
+    document.head.appendChild(script)
+    CustomJavascriptConstructorNode.cache[definedCode] = window[tempClassName]
+  }
+
+  protected _getCustomConstructor(): types.RunTimeNodeConstructor {
+    return this.isNodeJs() ? this._getNodeJsConstructor() : this._getBrowserConstructor()
+  }
+}
+
 class GrammarCustomConstructorsNode extends TreeNode {
   getKeywordMap() {
     const map = {}
     map[GrammarConstants.constructorNodeJs] = CustomNodeJsConstructorNode
     map[GrammarConstants.constructorBrowser] = CustomBrowserConstructorNode
+    map[GrammarConstants.constructorJavascript] = CustomJavascriptConstructorNode
     return map
   }
 
   getConstructorForEnvironment(): AbstractCustomConstructorNode {
+    const jsConstructor = this.getNode(GrammarConstants.constructorJavascript)
+    if (jsConstructor) return jsConstructor
     return this.getNode(this.isNodeJs() ? GrammarConstants.constructorNodeJs : GrammarConstants.constructorBrowser)
   }
 }