+++ releaseNotes.md
@@ -1,12 +1,13 @@
 edge / 2019-05-01
 =================
 - New: grammar name keyword
-- New: makeGraphSort function
+- New: makeGraphSort function. Also now used in grammar file prettification
 - Breaking: removed ability in grammar files to have a wordType and keyword share the same name
 - Breaking: getGraph now requires a uniqueId column. Throws if you attempt to extend a non-unique id
 - Breaking: instead of "grammar grammarName" oneliner now use the grammar name keyword
-- Fix: all word types now have default refex of [^ ]* so no need to specify it
+- Fix: all word types now have default regex of [^ ]* so no need to specify it
 - Fix: grammar code cleanup
+- Fix: small fixes to grammar sandbox
 
 20.0.0 / 2019-04-30
 ===================